pyproject.toml
[project]
name = "so101-playground"
version = "0.1.0"
description = "SO-101 follower robot arm playground and testing scripts"
authors = [
    {name = "gota", email = "gota@example.com"},
]
dependencies = [
    "feetech-servo-sdk>=1.0.0",
    "lerobot",
    "numpy",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "torchaudio>=2.4.0",
    "pytorch-triton-rocm>=3.2.0",
    "opencv-python>=4.8.0",
    "hydra-core>=1.3.0",
    "scipy>=1.15.3",
    "timm>=1.0.22",
    "natsort>=8.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.uv.sources]
torch = [
    { index = "pytorch-rocm" },
]
torchvision = [
    { index = "pytorch-rocm" },
]
torchaudio = [
    { index = "pytorch-rocm" },
]
pytorch-triton-rocm = [
    { index = "pytorch-rocm" },
]
lerobot = { path = "../lerobot", editable = true }

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true
